Initial commit
Please Note I have pushed everything needed for a working build of a
test file. I am to tired to figure out which bits of this should be
ignored. feel free to do some cleaning. will be new pushes on a sub
branch though so be sure to tell me what needs to go.